+++ src/main/java/baseball/Application.java
@@ -2,6 +2,6 @@ package baseball;
 
 public class Application {
     public static void main(String[] args) {
-        BaseBallGame.getInstance().run();
+        BaseballGame.getInstance().run();
     }
 }
