+++ src/main/java/baseball/util/ErrorMessage.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.util;
 
 public class ErrorMessage {
     public static final String ERROR_WRONG_USER_NUMBER = "1부터 9사이의 숫자로 이루어진, 중복되지 않는 숫자 세자리를 입력하세요.";
